updated cutoff.sh to run .cell files with different cutoff energies and plot the data

+++ src/cutoff.sh
@@ -1,18 +1,32 @@
 #!/bin/bash
 
-rm ../cutoff.dat
+old="$(pwd)"
+current="$(dirname "$0")"
+root=..
+cd $current
+cell=$root/cell
+data=$root/dat
+rm $data/cutoff.dat
 
+if ! [ -s $data/cutoff.dat ] ; then
 for i in {250..400..10}
 do
-file2=../cutoff.castep
+file2=$cell/cutoff.castep
 rm $file2
-file=../cutoff.param
+file=$cell/cutoff.param
 sed -i "/cut_off_energy/c\cut_off_energy     : "$i" eV." $file
 echo 'calculating energy using a cutoff energy of' $i
-mpirun -np 1 ../castep.mpi ../cutoff
+mpirun -np 1 $cell/castep.mpi $cell/cutoff
 
-STRING=$(grep "Total energy corrected for finite basis set" $file2)
-ENERGY=$(echo $STRING | awk '{print $9}')
-echo $i $ENERGY >> ../cutoff.dat
+energy=$(exec ./findenergy.sh $file2)
+ 
+echo $i $energy >> $data/cutoff.dat
 
 done
+cd $old
+
+else 
+echo "data file aready there!"
+fi
+
+./plot.sh $data/cutoff.dat 4
